se cambio el tipo de dato de los precio de los modelos de competencia y inscripcion a bigdecimal, ademas de poner como unique el nombre de la competencia. en el torneomodel se cambio el tipo de la fecha para solucionar una incosistencia en el crud del torneo y se agrego al repository de la competencia la funcion de busquepornombre

+++ Poo2025/src/main/java/ar/edu/unnoba/poo2025/torneos/model/InscripcionModel.java
@@ -1,5 +1,6 @@
 package ar.edu.unnoba.poo2025.torneos.model;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 import com.fasterxml.jackson.annotation.JsonBackReference;
@@ -29,8 +30,8 @@ public class InscripcionModel {
     private Date fechaInscripcion;
 
     @Column(nullable = false)
-    private double precioPagado;
-
+    private BigDecimal precioPagado;
+  
     @ManyToOne
     @JoinColumn(name = "id_participante", nullable = false)
     private ParticipanteModel participante;
